SLF4J conversion
git-svn-id: https://ha-jdbc.svn.sourceforge.net/svnroot/ha-jdbc/trunk@905 882a69a1-a10d-0410-8702-ae3607fe02de

+++ src/net/sf/hajdbc/balancer/RoundRobinBalancer.java
@@ -57,7 +57,7 @@ public class RoundRobinBalancer extends AbstractBalancer
 		
 		if (added)
 		{
-			int weight = database.getWeight().intValue();
+			int weight = database.getWeight();
 			
 			for (int i = 0; i < weight; ++i)
 			{
@@ -78,7 +78,7 @@ public class RoundRobinBalancer extends AbstractBalancer
 
 		if (removed)
 		{
-			int weight = database.getWeight().intValue();
+			int weight = database.getWeight();
 
 			for (int i = 0; i < weight; ++i)
 			{
